Feature: API: Add pcmk_api_init.
This function should be called by any external users of the libpacemaker
API before making any other calls.  For the moment, all it does is set
up libxml the way pacemaker expects, but more stuff could be added to it
if necessary.

Converting booth to use libpacemaker brought this problem up - if
crm_xml_init is not called prior to using libxml, libpacemaker will
crash.  However, crm_xml_init is deprecated and the replacement
(pcmk__xml_init) is a private function.