Fix: tools: Don't double-free XML in crm_verify after schema update
Currently, if an updated schema is acceptable, the old CIB XML is freed
by pcmk_update_configured_schema() and then again by
crm_verify.c:main(). Here we fix this by passing a pointer-to-pointer
into pcmk__verify(), so that crm_verify gets a pointer to the new XML
object instead of to the freed one.

This fixes a regression introduced by a744f1a8.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>