Feature: sysconfig: drop support for PCMK_dh_min_bits environment variable
This relied on a GnuTLS function that has been deprecated since the 2013 3.1.7
release, which we now require as a minimum.

The variable was an obscure setting that would allow TLS communication between
a client and server where one of them was really, really old. This will break
the ability to use Pacemaker Remote or remote CIB administration on a client or
server built with an ancient TLS library.