Feature: xml: Allow previously disallowed "required" op meta-attribute
This meta-attribute was disallowed beginning with schema version 3.0, to
avoid confusion.

As of the 2.x schemas, it was allowed and affected Pacemaker behavior.
Support for it was dropped when we introduced the 3.0 schemas. We
disallowed the meta-attribute in 3.x schemas so that users would not
mistakenly configure it expecting to affect behavior.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>