Feature: xml: Allow previously disallowed primitive meta-attributes
The same changes apply to resource templates as well.

These meta-attributes were disallowed beginning with schema version 3.0,
to avoid confusion.

As of the 2.x schemas, they were allowed and affected Pacemaker
behavior. Support for these meta-attributes was dropped when we
introduced the 3.0 schemas. We disallowed these properties in 3.x
schemas so that users would not mistakenly configure them expecting to
affect behavior.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>