Feature: various: Drop support for "#default" as an nvpair value
Special handling for "#default" as an nvpair values was first added by
commit ea1359b. It seems that it was never documented except as an
allowed value for the target-role meta-attribute. In that use case,
it's now already considered an invalid value.

The documentation for "target-role=#default" was only in the crm.dtd
file (not in the doc books), which was removed in 2.0.0 by commit
79a739c.

Thanks to populate_hash(), "#default" has been handled specially for ALL
nvpairs, not just a subset like resource meta-attributes. As a result,
we drop all nvpairs containing value="#default" during a schema upgrade
and disallow it as a value in the 4.0 schema.

Closes T742

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>