Refactor: CTS: Add an AclTest class.
A bunch of the ACL tests require some very specialized setup that
involve creating a temporary CIB as a certain user, modifying that CIB
in some way, and then running the test using that CIB as a different
user.

This is difficult to work into the existing scheme given that we want
the temporary CIB to go away when we're done with it, but also want to
be able to modify it and pass it around as a command line parameter.
Hence, a new class.  I'm not convinced this is a great name for it.