Fix: libpe_status: Use first action meta-attr within nvset if duplicates
Currently, if there are multiple meta-attributes with the same name in
the same meta_attributes element for a given action, or in a group of
meta_attributes elements with equal scores, the last one is applied.

With this commit, the first one is applied instead. This brings
meta-attribute processing into alignment with how option blocks
everywhere else behave.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>