Refactor: libpacemaker: apply_coloc_score() returns priority delta
...of the dependent. In other words, the apply_coloc_score() method now
returns the score that it added to the priv->priority member of its
dependent argument (which may be 0).

apply_coloc_score() affects priority only when determining clone
instance roles, not during assignment.

Nothing meaningfully uses this return value yet, but it will allow us to
fix a bug with colocations involving cloned groups.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>