Fix: libpacemaker: Apply promotion priority to cloned group instance
Rather than only to its members (only its first member in the case of
colocated groups). We sort instances for promotion based on each
instance's priority. In the case of a cloned group, each instance is a
group, and we call pcmk__apply_coloc_to_priority() for the group's
member(s) (first member for colocated groups, or all members for
non-colocated groups). The members' priorities are not visible to the
instance sorting code.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>