Refactor: libcrmcommon: consolidate create_request() "from" arguments
create_request() adds a sender system to the new request. Most callers passed
only the sys_from argument, which would be used as-is. If the caller passed
both sys_from and uuid_from, create_request() would combine the two.

Now, have the callers create the sender ID when needed. This simplifies a
confusing function signature and eliminates an unnecessary string duplication
in the common case.

This breaks API backward compatibility (but the interfaces will be dropped
shortly anyway).