Refactor: all: drop enum pcmk__cluster_msg
Use enum pcmk_ipc_server, which serves the same purpose for IPC, instead.
This makes sense since the message functions manipulate both IPC and cluster
messages.

The naming of pcmk_ipc_server is now unfortunate (pcmk_server would be more
appropriate now), but it's public API, and not worth the hassle of deprecating
it and defining a new enum with all the function/struct signature changes that
would be needed.