Refactor: daemons: always connect to CIB as crm_system_name
For native connections, this is used as PCMK__XA_CIB_CLIENTNAME, which is
passed to cib__client_triggers_refresh() and crm_is_daemon_name(), and
otherwise used only for logging. For these purposes, the change is
insignificant. (Or will be in a few commits ... with this commit alone, the
CIB won't recognize pacemaker-fenced as a daemon name.)

Also ensure that CIB signon() implementations can handle a NULL name argument.