Feature: schemas: Disallow nvpair name with id-ref attribute
This feature was added by commit 3912538 but is undocumented and
inconsistently implemented. The code often ignores name if there is an
id-ref, or in some places assumes that id and value exist if name
exists.

It could be useful in some circumstances, but it complicates XSL
transformations and it doesn't seem like we're losing much by dropping
it.

A previous commit already drops these during the 3-to-4 schema
transformation, replacing them with new nvpair elements.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>