Feature: various: Drop support for lifetime elements in constraints
When the CIB is upgraded to schema version 4.0, existing lifetime
elements in location constraints are converted to rules so that behavior
is preserved. However, existing lifetime elements in colocation and
order constraints are simply dropped, so behavior does change in that
case. Those constraint types don't support rules and there are no
current plans to add support.

Closes T271

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>