Feature: schemas: Disallow nvpairs with value "#default"
Previously, an nvpair with value "#default" would unset any value that
had been previously set for that option. Values that were processed even
later could then be applied.

The "#default" handling code was added by commit ea1359b. It seems that
it was never documented except as an allowed value for the target-role
meta-attribute. In that use case, it has been considered an invalid
value for a while.

The documentation for "target-role=#default" was only in the crm.dtd
file (not in the doc books), which was removed in 2.0.0 by commit
79a739c.

Ref T742

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>