Low: controller,libpacemaker: transition graph IDs should be positive
Previously, -1 was used to indicate an empty graph (an allocated graph object
not parsed from XML), and 0 was the first actual graph ID. However that led to
awkward log messages referring to "Transition -1" and "Transition 0".

Now, use 0 to indicate an empty graph, and 1 as the first actual ID. This still
leaves "Transition 0" for empty graphs (such as aborting a transition before
any transition has happened), but that can be addressed separately.