Feature: CIB: treat misconfigured date specification as not passing
Previously, an date_spec with a missing id would be ignored, but the date
expression would continue to be evaluated. Also, if a date_spec had an
attribute with an invalid range, that attribute would be ignored, but the rest
of the date_spec would continue to be evaluated.

Now, these are treated as configuration errors, and the entire date expression
fails.