Feature: CIB: treat misconfigured duration as not passing
Previously, a duration with a missing id would still be evaluated, and a
duration with an invalid attribute would have that attribute ignored but
otherwise continue to be evaluated. Now, these are treated as configuration
errors, and the entire date expression fails.

Also, the range_valid_start_duration_missing_id unit test previously had a
wrong test input, which has been fixed.