Feature: CIB: treat misconfigured attribute expression as not passing
Previously, an attribute expression with a missing id, invalid value-source,
value missing or present when inappropriate, or invalid type would have that
part ignored or defaulted, and the rest of the attribute expression would
continue to be evaluated.

Now, these are treated as configuration errors, and the entire attribute
expression fails.