Feature: various: Drop support for can_fail operation meta-attribute
It is undocumented and deprecated (with a warning).

It seems to be intended to allow a transition to continue and complete
successfully if an action with can_fail="true" fails. In practice, the
attribute manager updates the fail count, which aborts the transition
anyway.

Further, this option takes effect only while running a transition.
on-fail="ignore", which takes effect while processing resource history,
is thus not a replacement.

In light of all of the above, we simply drop this option if it's set.

Closes T268

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>