Feature: include: Add G_LOG_DOMAIN to crm_internal.h.
The glib docs state that this should not be defined in a public header
file.  The reason for this is to make sure nothing that links against
your code can see your definition of G_LOG_DOMAIN.  I think we're okay
here - the only places that include crm_internal.h are source files or
headers that would not be shipped (like crm_mon.h).

Related: T837