Test: cts-fencing: fix no-nodeid test
The test assumed the dummy fence agent did not support the nodeid
parameter, but it does. The test worked before non-clustered mode was
dropped because "localhost" would be used as the local node name, which
would not have a node ID.

Now, create another dummy alias fence_dummy_no_nodeid that strips out
nodeid from metadata, and use that in the test.