Low: scheduler: create local node after unpacking everything
Previously, the scheduler create a fake node for the local host name if
one didn't already exist after unpacking the nodes section. However, a
Pacemaker Remote node can be created after unpacking the resources and
status sections, so move the creation after all that.

This fixes a regression introduced by fd8fb8677a (not in a release)
where crm_mon on a Pacemaker Remote node showed the local node as an
offline cluster node.