Feature: CIB: ignore location constraints if score attribute is unset or invalid
If a location constraint has a rule with score-attribute, the constraint
score on a given node is taken from the value of the specified node
attribute on that node.

Originally, if the score attribute were unset, the score would be
treated as -INFINITY, and if the score attribute were invalid, the score
would be treated as 0.

Now, simply ignore the location constraint for an unset or invalid score
attribute.

This breaks behavioral backward compatibility.