Feature: schemas: Disallow role_after_failure resource meta-attribute
It is undocumented and deprecated (with a warning) since 2.0.0. It was
added by commit ea1359b3.

An XSL transformation now drops the meta-attribute if set. This may
change behavior. Preserving behavior via the on-fail operation
meta-attribute is not feasible for all cases. (For example, there is no
on-fail value that would lead to the Promoted role after failure.)

Ref T272

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>