High: libcrmcommon: Don't assert on failure to write errors.
Only text output mode is affected by this - other output formats write
to a buffer, so an error there is likely a memory problem and should be
asserted on.  vprintf/fprintf can return a negative value if the file
handle is closed (say, with shell output redirection) and we don't want
to assert on that.

Fixes T891