API: libstonithd: Add stonith_validate.
This is the guts of the existing stonith_api_validate function broken
out into a new function that takes a GHashTable as a parameter instead
of a list of stonith_key_value_t.  The old stonith_api_validate then
just wraps this new function and changes the return value and parameter
types to keep API compatibility.