API: libpacemaker: Take a GList * instead of a stonith_key_value_t *.
We are likely to get rid of the latter type at some point in the future,
and even if we don't, it shouldn't be used as the argument to a public
API function.

However, st->cmds->register_level_full still requires that as a
parameter and we can't change it.  So, take a GList everywhere and
convert the type at the last minute before calling the stonith function.