API: libpacemaker: Take a GList * instead of a stonith_key_value_t *.
Just like with pcmk_fence_register_level, pcmk_fence_validate should
not take a type we're likely to be getting rid of.  Also just like with
that function, we are still in the situation where
st->cmds->register_device wants the old type so we need to convert just
before calling.