Low: cts: Don't include the digest in expected output.
We are seeing a problem during test builds where the digest can differ.
This problem has not been seen when doing builds locally with mock, nor
on the clusterlabs CI infrastructure.  It additionally does not seem to
be a symptom of any other problem, since we've had no reports with weird
CIB differing problems.

Note that the digest was not included in the expected output from the
moment it was first committed in 2008 until I converted cts-cli.in to
python.  Presumedly, this was originally done because the input CIB
would be changing and they didn't also want to have to remember to
change the expected output.  Unfortunately, this makes it difficult to
know if we would have been seeing these digest differences all along.