Refactor: daemons: Get rid of an implicit fallthrough.
Support for __attribute__((fallthrough)) varies between compilers and
releases, so we've got macros to detect and deal with it.  However,
there's only one spot we are making use of the implicit fallthrough and
it only saves duplication of one line.  Plus, that's really kind of a
misfeature of C (hence the compiler warning).

So, just get rid of it and remember to not use implicit fallthroughs in
the future.

Fixes T558