Low: tools: crm_simulate should error on too-new feature sets.
The error message is hidden and only gets displayed if -V is given on
the command line.  Adding config error/warning handlers will cause the
error to be displayed regardless.

This could have been implemented in a couple ways, and there's tradeoffs
here.  I've chosen to duplicate what's happening in crm_verify, but
instead of checking for verbosity (which is a global variable in that
file), I'm checking out->is_quiet.

This means that if you do `crm_simulate -Q`, you won't see the error
message but you will get an error return code.  This also means that
`crm_simulate -Q -VVVV...`, you still won't see the error message.  This
may be a bug, but I'm not sure who would do that and I also think these
sorts of problems are pervasive in our command line tools.

Fix T521