Low: tools: crm_ticket should error on too-new feature sets.
This is just like the previous patch to crm_simulate.  However, one
additional problem here is that it relies on using the deprecated -Q
command line option.  On the other hand, I think this is okay because we
have a lot of work to do straightening out these sorts of options for
all our command line tools.  This is just one more thing we'll have to
deal with at that time.