Low: tools, libpacemaker: Check the return value of pcmk__schedule_actions.
This takes care of most callers - the ones in the daemons are unlikely
to be a problem.  This allows catching the too-new schema condition in
various other tools and displaying an error message to the user.

Note that a couple other callers don't need to check the return value.
I've added comments explaining why.