Refactor: libpacemaker: Don't continue if pcmk_unpack_scheduler_input fails.
This function can return a couple error codes, most notably when called
on input with a feature set that is newer than the latest supported.  In
that case, the caller should return its own error instead f trying to
continue on with an unpopulated scheduler object.  This prevents a
cascade of error messages.