Refactor: libcrmcommon: Use our own logging for gnutls.
Note that the gnutls docs mention that larger logging levels can
potentially reveal sensitive information.  However, what we are doing
here should be okay because it would require the administrator to enable
pacemaker logging at the debugging level and if that's happening, they
are probably trying to debug a problem that could be helped by the
increased gnutls logging.