Refactor: libcrmcommon: pcmk__x509_enabled doesn't need an arg.
For Pacemaker Remote connections, both ends are daemons and will
therefore read the TLS environment variables out of the sysconfig file.
However, one end is a client and one end is a server.  So using that to
decide where to read the variables is not helpful.

Instead, just try both locations regardless of which end of the
connection we're checking.  If the PCMK_ version of a variable isn't
set, try the CIB_ version instead.