Feature: daemons: Enable TLS support for Pacemaker Remote nodes.
If the right environment variables are set in the sysconfig file,
enable TLS certificates on the server side of a Pacemaker Remote node
connection.  This works exactly like the remote CIB operation support
did, except here the remote node is the server and the cluster is the
client.  Name your certs appropriately.

Ref T365