Test: cts-lab: log and skip "crm_node -q" nonzero exit status
If the lab happens to try to run "crm_node -q" on a node and it can't
connect to the controller, it will get exit status 102
(CRM_EX_DISCONNECT) and no output.

Previously, this would result in an exception that stopped the lab.
Now, it is logged and the lab continues.