Refactor: libcrmcommon: Use gnutls_transport_set_int()
And gnutls_transport_get_int(). As of GnuTLS 3.1.9, we no longer have to
use the _ptr() functions. The _int() functions are a simpler
alternative, assuming TCP is the transport layer.

As of this commit, there's no longer any real benefit to keeping
pcmk__tls_get_client_sock(), except to provide a layer of abstraction so
that GnuTLS library functions are called in fewer places.

https://gnutls.org/manual/html_node/Setting-up-the-transport-layer.html

Ref T967

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>