Refactor: daemons: Unindent a block of code in action_complete.
First, invert the test to see if this is a systemd resource.  If it's
not, there's no way goagain was set so we can just bail out to the end
of the function.

Then having done that, everything that was in the block can be
unindented.

No other code changes.