Feature: libcrmservices: Add a systemd job complete callback.
This will get called when we receive a systemd message from dbus and it
makes it through the filter in filter_systemd_signals.  At the moment,
this is only the JobRemoved signal which can be used to determine if a
start action we initiated has started.

This could be expanded to other signals in the future, if that would be
useful.