Feature: daemons: Don't handle start actions in action_complete.
The action_complete function is called when a systemd action finishes.
Unfortunately, StartUnit actions finish and return a value immediately,
not when the unit we asked it to start is actually up and running.

For this reason, action_complete handles start actions by then
scheduling a follow-up monitor action to poll the service and see if
it's actually up and running yet.  It then takes the appropriate action
when this happens.

Since we're moving towards waiting for systemd to tells us a service is
running, we don't need to do any of this anymore.  However, stop actions
and monitors still need this code.  For the moment, it's easiest to just
bail out for start actions.