Refactor: libservices: Add the systemd unit name to opaque data.
Internally, pacemaker uses a shorter resource-based name when referring
to systemd actions.  However, systemd itself wants a real unit name so
we append ".service" or whatever to the name at exec time.  Because we
do this at the last minute and don't save it anywhere, there's no way to
figure out what action a systemd unit is associated with later.

So, add that to the opaque data.  This also requires an accessor
function since the opaque data is supposed to be private.