Refactor: libcrmservices: Add a function to get an inflight op.
An inflight operation is one that is currently being executed - either
by us (in the case of something that we forked and execed), or systemd
(in the case of something we told it to start up).  Here, we add a
function that just get a systemd inflight op that matches a given unit
name.